refactor: delete useless types (Session), fix minor bugs
